Create a gp_options Object

Constructs a gp_options object with configurable parameters. The parameters can be set from the defaults, from R options, or directly when the options object is created.

Usage

gp_options(
  ...,
  pin_directory = getOption("gitpins.pin_directory", here::here("gitpins"))
)

Arguments

...

Reserved. All arguments must be named.

pin_directory

A character string specifying the directory to use for the pin directory (the local git repository where downloaded resources are stored and versioned).

Value

A list representing the gp_options object.